+++ regulations/tests/views_utils_test.py
@@ -1,10 +1,10 @@
-#vim: set encoding=utf-8
+# vim: set encoding=utf-8
 from unittest import TestCase
 from mock import patch
 
 from django.conf import settings
 
-from regulations.views.utils import *
+from regulations.views import utils
 
 
 class UtilsTest(TestCase):
@@ -24,7 +24,7 @@ class UtilsTest(TestCase):
 
     def test_get_layer_list(self):
         names = 'meta,meta,GRAPHICS,fakelayer,internal'
-        layer_list = get_layer_list(names)
+        layer_list = utils.get_layer_list(names)
         self.assertEquals(set(['meta', 'internal', 'graphics']), layer_list)
 
     @patch('regulations.generator.generator.LayerCreator.get_layer_json')
@@ -32,22 +32,22 @@ class UtilsTest(TestCase):
         get_layer_json.return_value = {'layer': 'layer'}
 
         layer_names = 'graphics,meta'
-        appliers = handle_specified_layers(layer_names, '205', '2013-1')
+        appliers = utils.handle_specified_layers(layer_names, '205', '2013-1')
         self.assertEquals(3, len(appliers))
 
     def test_add_extras_env(self):
         context = {}
 
         settings.JS_DEBUG = True
-        add_extras(context)
+        utils.add_extras(context)
         self.assertEqual('source', context['env'])
 
         settings.JS_DEBUG = False
-        add_extras(context)
+        utils.add_extras(context)
         self.assertEqual('built', context['env'])
 
         del(settings.JS_DEBUG)
-        add_extras(context)
+        utils.add_extras(context)
         self.assertEqual('built', context['env'])
 
     def test_add_extras(self):
@@ -63,21 +63,33 @@ class UtilsTest(TestCase):
             }
         }
 
-        add_extras(context)
+        utils.add_extras(context)
 
         self.assertTrue('APP_PREFIX' in context)
         self.assertTrue('env' in context)
 
-        self.assertEquals('gtm-site-id', context['ANALYTICS']['GOOGLE']['GTM_SITE_ID'])
-        self.assertEquals('ga-site-id', context['ANALYTICS']['GOOGLE']['GA_SITE_ID'])
+        self.assertEquals('gtm-site-id',
+                          context['ANALYTICS']['GOOGLE']['GTM_SITE_ID'])
+        self.assertEquals('ga-site-id',
+                          context['ANALYTICS']['GOOGLE']['GA_SITE_ID'])
         self.assertEquals('agency', context['ANALYTICS']['DAP']['AGENCY'])
-        self.assertEquals('sub-agency', context['ANALYTICS']['DAP']['SUBAGENCY'])
-        self.assertEquals('agency=agency&subagency=sub-agency', context['ANALYTICS']['DAP']['DAP_URL_PARAMS'] )
+        self.assertEquals('sub-agency',
+                          context['ANALYTICS']['DAP']['SUBAGENCY'])
+        self.assertEquals('agency=agency&subagency=sub-agency',
+                          context['ANALYTICS']['DAP']['DAP_URL_PARAMS'])
 
     @patch('regulations.views.utils.fetch_toc')
     def test_first_section(self, fetch_toc):
         fetch_toc.return_value = [
             {'section_id': '204-100', 'index': ['204', '100']},
             {'section_id': '204-101', 'index': ['204', '101']}]
-        first = first_section('204', '2')
+        first = utils.first_section('204', '2')
         self.assertEqual(first, '204-100')
+
+    def test_make_sortable(self):
+        """Verify that strings get decomposed correctly into sortable tuples"""
+        self.assertEqual(utils.make_sortable("abc"), ("abc",))
+        self.assertEqual(utils.make_sortable("123"), (123,))
+        self.assertEqual(utils.make_sortable("abc123def456"),
+                         ("abc", 123, "def", 456))
+        self.assertEqual(utils.make_sortable("123abc456"), (123, "abc", 456))
